Summary: DMA buffer caching fix, RCC periodic worker fix, VHDL raw property support, zero width data paths adapt/os/ocpios/interfaces/include/KernelDriver.h: better comment adapt/os/ocpios/linux/driver/opencpi.c: use correct enum for cacheability components/Makefile: add file_write.hdl components/file_write.hdl/Makefile: restrict building this to sim targets core/container/hdl/src/PciDriver.cxx: initialize caching enum for kernel driver requests core/container/hdl/src/ocpihdl_main.cxx: remove redundant lines in usage message core/container/rcc/src/OcpiRccWorker.cxx: run condition timer with zero port masks fixed core/dataplane/rdma_drivers/pci/src/DtPCIPioXfer.cxx: initialize cacheability enum for kernel driver core/local/util/property/include/OcpiUtilDataTypes.h: comment generic_env_post.sh: comment hdl/assemblies/testbias2/testbias2_cont.xml: add newline at eof hdl/devices/adc.rcc/adc.c: use size_t for memories array hdl/platforms/Makefile: add package name since this serves as a sort of lib for platform workers hdl/platforms/common/mkCTop4B.v: enable dac interface - it should be pacified in the app, not here hdl/primitives/ocpi/control_decoder.vhd: remove unused signal hdl/primitives/ocpi/decoder.vhd: enable raw property access hdl/primitives/ocpi/master.vhd: fix >8  byte sizes hdl/primitives/ocpi/slave.vhd: ditto hdl/primitives/ocpi/types_body.vhd: add convenience to_string from VHDL string type hdl/primitives/ocpi/types_pkg.vhd: ditto hdl/primitives/ocpi/wci_pkg.vhd: enable raw property access jkvm6.sh: update to ISE 14.6 test/hdl/scripts/loadBitStream: better message test/hdl/scripts/probeXilinxUsb: remove previous log files before running impact tools/cdk/cdkutils/src/cdkutils.cxx: indent tools/cdk/examples/bestfit_test/bestfit.cxx: size_t for getBuffer tools/cdk/examples/canny/canny.cxx: ditto tools/cdk/examples/cat_xml/catHello.cxx: ditto tools/cdk/examples/ext2file_xml/ext2file.cxx: ditto tools/cdk/examples/feature_detection/feature_detection.cxx: ditto tools/cdk/examples/hello/hello.cxx: ditto tools/cdk/examples/hello_xml/appHello.cxx: ditto tools/cdk/examples/image_filtering/image_filtering.cxx: ditto tools/cdk/examples/optical_flow/optical_flow.cxx: ditto tools/cdk/examples/optical_flow_xml/optical_flow.cxx: ditto tools/cdk/include/hdl/hdl-assy.mk: UUID should depend on the platform tools/cdk/include/hdl/hdl-worker.mk: hdl language should not be case sensitive tools/cdk/include/hdl/xst.mk: include "work" library in lso file in more cases tools/cdk/include/lib.mk: put package processing in its own file tools/cdk/include/util.mk: work around "realpath" bug with nfs mounts, add our own OcpiAbsPath tools/cdk/ocpigen/bin/src/ocpigen_main.cxx: cleanup with Worker methods, preliminary new container gen tools/cdk/ocpigen/ocpigen/include/wip.h: new signal info, raw properties, Worker methods tools/cdk/ocpigen/ocpigen/src/wip_HDL.cxx: worker methods, raw properties, sequence lengths, signals -- also zero length data paths tools/cdk/ocpigen/ocpigen/src/wip_parse.cxx: worker methods, raw properties, case insensitive prop names -- also new signal parser

+++ hdl/devices/adc.rcc/adc.c
@@ -50,7 +50,7 @@
 
 #define FILE_POINTER_INDEX 0
 
-static uint32_t memories [ ] =
+static size_t memories [ ] =
 {
   sizeof ( FILE* ), /* Input data file pointer */
   0
